pyproject.toml
[project]
name = "uipath-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckduckgo-search>=8.0.2",
    "langchain-openai>=0.3.17",
    "langchain-pinecone>=0.2.6",
    "pinecone>=6.0.2",
    "streamlit>=1.45.1",
    "uipath-langchain>=0.0.110",
]
